Lab4/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
bidict==0.22.1
blinker==1.6.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.6
coloredlogs==15.0.1
contourpy==1.3.0
crispy-bootstrap5==2024.2
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==5.1.1
django-bootstrap5==23.4
django-crispy-forms==2.3
django-debug-toolbar==4.2.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==2.3.2
Flask-Login==0.6.2
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.0.5
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gdown==5.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
humanfriendly==10.0
idna==3.8
imageio==2.35.1
imgviz==1.7.5
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.5.0
joblib==1.2.0
kafka-python==2.0.2
kiwisolver==1.4.5
labelme @ git+https://github.com/labelmeai/labelme@ecc83c2771c6c2f78649adad3a2220caf3e72d93
lazy_loader==0.4
loguru==0.7.2
markdown2==2.5.0
MarkupSafe==2.1.3
matplotlib==3.9.2
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.2.4
natsort==8.4.0
networkx==3.3
numpy==1.24.3
onnxruntime==1.19.2
openai==1.44.0
opencv-python==4.10.0.84
osam==0.2.2
packaging==24.1
pandas==2.0.1
pillow==10.4.0
pipenv==2023.11.15
platformdirs==4.1.0
protobuf==5.28.0
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.9.0
pydantic_core==2.23.2
pygame==2.5.0
PyMySQL==1.1.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0.2
QtPy==2.4.1
requests==2.31.0
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.20
sqlparse==0.5.1
sympy==1.13.2
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.1.0
tifffile==2024.8.30
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.85
ultralytics-thop==2.0.6
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==2.3.7
